Auto Payment Instruction Creation Reports : Data Saving on Loan approval process for Reports Auto creation of Refunds Job Disbursement Download - Bank level formats CD Scheme level changes
git-svn-id: http://192.168.110.32/svn/Products/PFF/git-migration/branches/mlod@136342 6963f9ea-81fa-44da-aa10-7ebde50aa541

+++ PFSJava/src/main/java/com/pennant/backend/service/payment/impl/PaymentHeaderServiceImpl.java
@@ -89,6 +89,7 @@ import com.pennant.backend.util.DisbursementConstants;
 import com.pennant.backend.util.FinanceConstants;
 import com.pennant.backend.util.PennantConstants;
 import com.pennant.backend.util.PennantJavaUtil;
+import com.pennant.backend.util.UploadConstants;
 import com.pennant.cache.util.AccountingConfigCache;
 import com.pennanttech.pennapps.core.InterfaceException;
 import com.pennanttech.pennapps.core.model.ErrorDetail;
@@ -398,13 +399,14 @@ public class PaymentHeaderServiceImpl extends GenericService<PaymentHeader> impl
 			}
 		}
 
-		auditHeader.setAuditDetails(deleteChilds(paymentHeader, TableType.TEMP_TAB, auditHeader.getAuditTranType()));
-		String[] fields = PennantJavaUtil.getFieldDetails(new PaymentHeader(), paymentHeader.getExcludeFields());
-		auditHeader.setAuditDetail(new AuditDetail(auditHeader.getAuditTranType(), 1, fields[0], fields[1],
-				paymentHeader.getBefImage(), paymentHeader));
-		getAuditHeaderDAO().addAudit(auditHeader);
-
-		getPaymentHeaderDAO().delete(paymentHeader, TableType.TEMP_TAB);
+		if (!StringUtils.equals(UploadConstants.FINSOURCE_ID_CD_PAY_UPLOAD, paymentHeader.getFinSource())) {
+			auditHeader.setAuditDetails(deleteChilds(paymentHeader, TableType.TEMP_TAB, auditHeader.getAuditTranType()));
+			String[] fields = PennantJavaUtil.getFieldDetails(new PaymentHeader(), paymentHeader.getExcludeFields());
+			auditHeader.setAuditDetail(new AuditDetail(auditHeader.getAuditTranType(), 1, fields[0], fields[1],
+					paymentHeader.getBefImage(), paymentHeader));
+			getAuditHeaderDAO().addAudit(auditHeader);
+			getPaymentHeaderDAO().delete(paymentHeader, TableType.TEMP_TAB);
+		}
 		auditHeader.setAuditTranType(PennantConstants.TRAN_WF);
 		getAuditHeaderDAO().addAudit(auditHeader);
